refactor: DB Trans 오브젝트 경로 변경
