+++ src/main/java/com/backend/connectable/event/domain/repository/EventRepositoryCustom.java
@@ -1,6 +1,6 @@
 package com.backend.connectable.event.domain.repository;
 
-import com.backend.connectable.event.ui.dto.EventDetail;
+import com.backend.connectable.event.domain.dto.EventDetail;
 
 public interface EventRepositoryCustom {
